requirements.txt
requests
pandas
apache-airflow
apache-airflow-providers-postgres
beautifulsoup4



